requirements.txt
python-dotenv
flask
openai
requests